Move torrent_info, check_seed_limits to utils
